Add Table Of Contents capability to docset
adds table of contents to every page, with a custom mapping to Dash's
internal entry types. I left the map in place so that we can adjust the
type if needed later, if something else makes more sense.

adds custom css so that it does not need to be maintained as much.